Changing the components to a much dense level to work easily with props and state
