+++ src/components/Pages/my-event-component/sub-components/hosting.component.tsx
@@ -1,0 +1,47 @@
+import React from 'react'
+import { makeStyles } from '@material-ui/core/styles';
+import Table from '@material-ui/core/Table';
+import TableBody from '@material-ui/core/TableBody';
+import TableCell from '@material-ui/core/TableCell';
+import TableContainer from '@material-ui/core/TableContainer';
+import TableHead from '@material-ui/core/TableHead';
+import TableRow from '@material-ui/core/TableRow';
+import Paper from '@material-ui/core/Paper';
+
+const useStyles = makeStyles({
+    table: {
+      minWidth: 650,
+    },
+  });
+
+export const HostingComponent: React.FC = ()=>{
+    const classes = useStyles();
+    return(
+        <React.Fragment>
+            <TableContainer component={Paper}>
+                <Table className={classes.table} aria-label="simple table">
+                    <TableHead>
+                    <TableRow>
+                        <TableCell>Dessert (100g serving)</TableCell>
+                        <TableCell align="right">Calories</TableCell>
+                        <TableCell align="right">Fat&nbsp;(g)</TableCell>
+                        <TableCell align="right">Carbs&nbsp;(g)</TableCell>
+                        <TableCell align="right">Protein&nbsp;(g)</TableCell>
+                    </TableRow>
+                    </TableHead>
+                    <TableBody>
+                        <TableRow >
+                        <TableCell component="th" scope="row">
+                        5
+                        </TableCell>
+                        <TableCell align="right">1</TableCell>
+                        <TableCell align="right">2</TableCell>
+                        <TableCell align="right">3</TableCell>
+                        <TableCell align="right">4</TableCell>
+                        </TableRow>
+                    </TableBody>
+                </Table>
+            </TableContainer>
+        </React.Fragment>
+    )
+}
