--- src/components/Pages/my-event.component.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import { UserEventComponent } from "./users-event-container-component/user-event-container.component";
-
-
-
-export const MyEventComponent: React.FC = () => {
-  return (
-    <React.Fragment>
-      <UserEventComponent />
-    </React.Fragment>
-  );
-};
