--- src/components/sub-components/date-component.tsx
@@ -1,49 +1,0 @@
-import React from 'react'
-import { Grid } from "@material-ui/core";
-import {} from '@material-ui/icons';
-import DateFnsUtils from '@date-io/date-fns';
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker
-} from '@material-ui/pickers';
-
-
-export const DatePickerComponent: React.FC = () => {
-    
-    const [selectedDate, setSelectedDate] = React.useState<Date | null>(
-        new Date('2014-08-18T21:11:54'),
-      );
-    
-    const handleDateChange = (date: Date | null) => {
-        setSelectedDate(date);
-    };
-    
-    return (
-        <MuiPickersUtilsProvider utils={DateFnsUtils}>
-            <Grid container justify="space-around">
-                <KeyboardDatePicker
-                    margin="normal"
-                    id="date-picker-dialog"
-                    label="Event Starts On:"
-                    format="MM/dd/yyyy"
-                    value={selectedDate}
-                    onChange={handleDateChange}
-                    KeyboardButtonProps={{
-                        'aria-label': 'change date',
-                    }}
-                />
-                <KeyboardDatePicker
-                    margin="normal"
-                    id="date-picker-dialog"
-                    label="Event Ends On:"
-                    format="MM/dd/yyyy"
-                    value={selectedDate}
-                    onChange={handleDateChange}
-                    KeyboardButtonProps={{
-                        'aria-label': 'change date',
-                    }}
-                />
-            </Grid>
-      </MuiPickersUtilsProvider>
-    );
-};
