--- src/components/sub-components/image-component.tsx
@@ -1,44 +1,0 @@
-import React from 'react';
-import { makeStyles, createStyles, Theme, 
-    Button, IconButton } from '@material-ui/core';
-import { PhotoCamera } from '@material-ui/icons';
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      '& > *': {
-        margin: theme.spacing(1),
-      },
-    },
-    input: {
-      display: 'none',
-    },
-  }),
-);
-
-export const ImageUploadButton: React.FC = () => {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.root}>
-      <input
-        accept="image/*"
-        className={classes.input}
-        id="contained-button-file"
-        multiple
-        type="file"
-      />
-      <label htmlFor="contained-button-file">
-        <Button variant="contained" color="primary" component="span">
-          Upload
-        </Button>
-      </label>
-      <input accept="image/*" className={classes.input} id="icon-button-file" type="file" />
-      <label htmlFor="icon-button-file">
-        <IconButton color="primary" aria-label="upload picture" component="span">
-          <PhotoCamera />
-        </IconButton>
-      </label>
-    </div>
-  );
-};
