--- src/components/sub-components/redirect-modal-component.tsx
@@ -1,55 +1,0 @@
-import React, { useState } from 'react'
-import { Button, Slide, Dialog, DialogTitle, 
-    DialogContent, DialogContentText, DialogActions } from '@material-ui/core';
-import { TransitionProps } from '@material-ui/core/transitions/transition';
-
-const Transition = React.forwardRef(function Transition(
-    props: TransitionProps & { children?: React.ReactElement<any, any> },
-    ref: React.Ref<unknown>,
-  ) {
-    return <Slide direction="up" ref={ref} {...props} />;
-  });
-  
-  export const RedirectModalComponent: React.FC = () => {
-    const [open, setOpen] = useState(false);
-  
-    const handleClickOpen = () => {
-      setOpen(true);
-    };
-  
-    const handleClose = () => {
-      setOpen(false);
-    };
-  
-    return (
-      <div>
-        <Button variant="outlined" color="primary" onClick={handleClickOpen}>
-          Go to Event Forum
-        </Button>
-        <Dialog
-          open={open}
-          TransitionComponent={Transition}
-          keepMounted
-          onClose={handleClose}
-          aria-labelledby="alert-dialog-slide-title"
-          aria-describedby="alert-dialog-slide-description"
-        >
-          <DialogTitle id="alert-dialog-slide-title">{"You sure you want to leave?"}</DialogTitle>
-          <DialogContent>
-            <DialogContentText id="alert-dialog-slide-description">
-              Leaving this page will redirect you to the 
-              Event Forum Page. Happy Hunting!!
-            </DialogContentText>
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={handleClose} color="primary">
-              No
-            </Button>
-            <Button href="#" onClick={handleClose} color="primary">
-              Yes
-            </Button>
-          </DialogActions>
-        </Dialog>
-      </div>
-    );
-  }
